+++ src/main/java/com/solver/solver_be/domain/user/controller/UserController.java
@@ -1,9 +1,6 @@
 package com.solver.solver_be.domain.user.controller;
 
-import com.solver.solver_be.domain.user.dto.AdminSignupRequestDto;
-import com.solver.solver_be.domain.user.dto.GuestSignupRequestDto;
-import com.solver.solver_be.domain.user.dto.LoginRequestDto;
-import com.solver.solver_be.domain.user.dto.PasswordChangeRequestDto;
+import com.solver.solver_be.domain.user.dto.*;
 import com.solver.solver_be.domain.user.service.AdminService;
 import com.solver.solver_be.domain.user.service.GuestService;
 import com.solver.solver_be.global.response.GlobalResponseDto;
@@ -14,6 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
+import javax.mail.MessagingException;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
@@ -44,22 +42,46 @@ public class UserController {
         return adminService.changePassword(passwordChangeRequestDto, userDetails.getAdmin());
     }
 
-    // 5. Guest Signup
+    // 5. Found Admin userId
+    @PostMapping("/admin/search")
+    public ResponseEntity<GlobalResponseDto> adminIdSearch(@RequestBody UserSearchRequestDto userSearchRequestDto) throws MessagingException {
+        return adminService.findAdminSearchId(userSearchRequestDto);
+    }
+
+    // 6. Reset Admin Password
+    @PostMapping("/admin/password")
+    public ResponseEntity<GlobalResponseDto> resetAdminPassword(@RequestBody PasswordResetRequestDto passwordResetRequestDto) throws MessagingException {
+        return adminService.resetAdminPassword(passwordResetRequestDto);
+    }
+
+    // 7. Guest Signup
     @PostMapping("/signup/guest")
     public ResponseEntity<GlobalResponseDto> signupGuest(@Valid @RequestBody GuestSignupRequestDto signupRequestDto) {
         return guestService.signupGuest(signupRequestDto);
     }
 
-    // 6. Guest Login
+    // 8. Guest Login
     @PostMapping("/login/guest")
     public ResponseEntity<GlobalResponseDto> loginGuest(@RequestBody LoginRequestDto loginRequestDto, HttpServletResponse response) {
         return guestService.login(loginRequestDto, response);
     }
 
-    // 7. Guest change password
+    // 9. Guest change password
     @PutMapping("/change/password/guest")
     public ResponseEntity<GlobalResponseDto> changeGuest(@Valid @RequestBody PasswordChangeRequestDto passwordChangeRequestDto,
                                                             @AuthenticationPrincipal UserDetailsImpl userDetails){
         return guestService.changePassword(passwordChangeRequestDto, userDetails.getGuest());
+    }
+
+    // 10. Found Guest userId
+    @PostMapping("/guest/search")
+    public ResponseEntity<GlobalResponseDto> guestIdSearch(@RequestBody UserSearchRequestDto userSearchRequestDto) throws MessagingException {
+        return guestService.findGuestSearchId(userSearchRequestDto);
+    }
+
+    // 11. Reset Guest Password
+    @PostMapping("/guest/password")
+    public ResponseEntity<GlobalResponseDto> resetGuestPassword(@RequestBody PasswordResetRequestDto passwordResetRequestDto) throws MessagingException {
+        return guestService.resetGuestPassword(passwordResetRequestDto);
     }
 }
